feat(carto): refactor main carto
