+++ front/src/components/baseComponents/customSnackbar.component.tsx
@@ -10,6 +10,7 @@ interface Props {
   duration: number;
   visible: boolean;
   isOnTop?: boolean;
+  marginTopValue?: number;
   backgroundColor?: string;
   onDismiss: () => void;
   textColor?: string;
@@ -20,6 +21,7 @@ const CustomSnackbar: React.FC<Props> = ({
   duration,
   visible,
   isOnTop,
+  marginTopValue,
   backgroundColor,
   onDismiss,
   textColor,
@@ -39,6 +41,8 @@ const CustomSnackbar: React.FC<Props> = ({
     color: textColor ?? Colors.dark.text,
   };
 
+  const marginTop = { marginTop: marginTopValue };
+
   return (
     <>
       {visible && (
@@ -46,7 +50,9 @@ const CustomSnackbar: React.FC<Props> = ({
           animation="fadeIn"
           duration={200}
           style={[
-            isOnTop ? styles.snackbarViewTop : styles.snackbarViewBottom,
+            isOnTop
+              ? [styles.snackbarViewTop, marginTop]
+              : styles.snackbarViewBottom,
             styles.snackbarView,
             snackbarStyle,
           ]}
