+++ front/src/styles/spacings.ts
@@ -11,7 +11,7 @@ export const Margins = {
   largest: scaleNormalize(36),
   evenMoreLargest: scaleNormalize(50),
   step: scaleNormalize(60),
-  snackbarMargin: scaleNormalize(80),
+  simpleCartoSnackbarMargin: scaleNormalize(90),
 };
 
 export const Paddings = {
